refactor: create modules for app
