+++ src/database.ts
@@ -1,0 +1,17 @@
+import { Users } from 'types';
+
+export const users: Users = {
+  '8b056444-b6b3-46d4-acd9-416c119c5293': {
+    id: '8b056444-b6b3-46d4-acd9-416c119c5293',
+    username: 'Duxcoder',
+    age: '12',
+    hobbies: ['js'],
+  },
+
+  '75ba0f01-b739-4b63-a54e-b07d2241f796': {
+    id: '75ba0f01-b739-4b63-a54e-b07d2241f796',
+    username: 'Guest',
+    age: '98',
+    hobbies: ['cartoon'],
+  },
+};
