+++ src/utils.ts
@@ -1,0 +1,3 @@
+export const checkPathUser = (path: string) => {
+  return path.split('/').length === 4 && path.startsWith('/api/users/');
+};
